Move crypto operations for SSO nonce.
